Improved persistence and added scripts to automate startup.
A backup folder with everything important for development that usually sits in the volume folders is added. Sometimes Kafka may have corrupted logs and everything persisted is best deleted, so the backup folder provides a safety net.

Some more specific volumes have been added to improve persistence.

The powershell script for startup will now leave the terminal open after executed.

A new experimental powershell startup script has been added which does everything up to creating a topic. It is still unknown if there are any errors that could come from it, hence it is still experimental.

The termination script now waits for a key to be pressed to cose the terminal after taking down the kafka container.